Implement updates for better performance

+++ code/blockchain/contracts/Migrations.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 contract Migrations {
     address public owner;
-    uint public last_completed_migration;
+    uint256 public last_completed_migration;
 
     constructor() {
         owner = msg.sender;
@@ -17,7 +17,7 @@ contract Migrations {
         _;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         last_completed_migration = completed;
     }
